vxWorks: remove all code related to UNIX socket as it is not supported by vxWorks
